Fix warnings in NZXTHue1Controller.cpp

+++ Controllers/NZXTHue1Controller/NZXTHue1Controller.cpp
@@ -88,7 +88,7 @@ void NZXTHue1Controller::SetEffect
             \*-----------------------------------------------------*/
             for (std::size_t idx = 0; idx < 40; idx++)
             {
-                int pixel_idx = idx * 3;
+                int pixel_idx = (int)idx * 3;
                 RGBColor color = colors[color_idx];
                 color_data[pixel_idx + 0x00] = RGBGetGValue(color);
                 color_data[pixel_idx + 0x01] = RGBGetRValue(color);
@@ -111,7 +111,7 @@ void NZXTHue1Controller::SetEffect
         \*-----------------------------------------------------*/
         for (std::size_t idx = 0; idx < num_colors; idx++)
         {
-            int pixel_idx = idx * 3;
+            int pixel_idx = (int)idx * 3;
             RGBColor color = colors[idx];
             color_data[pixel_idx + 0x00] = RGBGetGValue(color);
             color_data[pixel_idx + 0x01] = RGBGetRValue(color);
@@ -138,7 +138,7 @@ void NZXTHue1Controller::SetLEDs
     \*-----------------------------------------------------*/
     for (std::size_t idx = 0; idx < num_colors; idx++)
     {
-        int pixel_idx = idx * 3;
+        int pixel_idx = (int)idx * 3;
         RGBColor color = colors[idx];
         color_data[pixel_idx + 0x00] = RGBGetGValue(color);
         color_data[pixel_idx + 0x01] = RGBGetRValue(color);
